autostart/autopoll doesn't exist in this driver anymore, so this is what claude suggested to replace it. It doesn't seem high value but its also not terrible, so I decided to keep it. I did the rename.
